Rename PresenterModel to HostModel
